+++ R/g.part6.R
@@ -106,18 +106,29 @@ g.part6 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
       skip = 0
     }
     if (params_general[["overwrite"]] == TRUE) skip = 0
-    cosinor_coef = NULL
+    Lnames = cosinor_coef = mdat = NULL
     if (skip == 0) {
       # Load time series:
       if (EXT == "RData") {
         load(file = paste0(metadatadir, "/meta/ms5.outraw/",
                            params_phyact[["part6_threshold_combi"]], "/", fnames.ms5raw[i]))
+        if (is.null(Lnames)) stop("Part 5 was processed with an older version of GGIR, reprocess part 5")
         mdat$time = mdat$timestamp # duplicate column because cosinor function expect columntime
       } else {
+        # use behavioural codes files to derive Lnames object (names of behavioural classes)
+        legendfile = dir(path = paste0(metadatadir, "/meta/ms5.outraw"), pattern = "behavioralcodes", full.names = TRUE)
+        if (length(legendfile) > 1) {
+          df <- file.info(legendfile)
+          legendfile = rownames(df)[which.max(df$mtime)]
+        } else if (length(legendfile) == 0) {
+          stop(paste0("behaviouralcodes file could not be found in ",
+                      paste0(metadatadir, "/meta/ms5.outraw")))
+        }
+        Lnames = data.table::fread(file = legendfile)$class_name
         mdat = data.table::fread(file = paste0(metadatadir, "/meta/ms5.outraw/", 
                                                params_phyact[["part6_threshold_combi"]],  "/", fnames.ms5raw[i]), data.table = FALSE)
       }
-      nfeatures = 50
+      nfeatures = 200
       summary = matrix(NA, nfeatures, 1)
       s_names = rep("", nfeatures)
       fi = 1
@@ -202,8 +213,9 @@ g.part6 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
       s_names[fi] = "N_valid_days"
       fi = fi + 1
       #=================================================================
-      # Cosinor analysis which comes with IV IS estimtes
+      # Circadian rhythm analysis
       if (do.cr == TRUE) {
+        # Cosinor analysis which comes with IV IS estimtes
         # Note: applyCosinorAnalyses below uses column invalidepoch to turn
         # imputed values to NA.
         colnames(ts)[which(colnames(ts) == "timenum")] = "time"
@@ -250,18 +262,31 @@ g.part6 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
                                  cosinor_coef$IVIS$phi)
         s_names[fi:(fi + 2)] = c("IS", "IV", "phi")
         fi = fi + 3
-      } else {
-        cosinor_coef = NULL
-        s_names[fi:(fi + 20)] = c("cosinor_timeOffsetHours", "cosinor_mes", 
-                                  "cosinor_amp", "cosinor_acrophase",
-                                  "cosinor_acrotime", "cosinor_ndays", "cosinor_R2", 
-                                  "cosinorExt_minimum", "cosinorExt_amp", 
-                                  "cosinorExt_alpha", "cosinorExt_beta", 
-                                  "cosinorExt_acrotime", "cosinorExt_UpMesor",
-                                  "cosinorExt_DownMesor", "cosinorExt_MESOR",
-                                  "cosinorExt_ndays", "cosinorExt_F_pseudo", 
-                                  "cosinorExt_R2", "IS", "IV", "phi")
-        fi = fi + 21
+        # Transition probabilities
+        for (fragmode in c("day", "spt")) {
+          ts_temp = ts
+          # turn other half of data to NA
+          ts_temp[which(ts$diur == ifelse(fragmode == "spt", 0, 1))] = NA
+          frag.out = g.fragmentation(frag.metrics = params_phyact[["frag.metrics"]],
+                                     LEVELS = ts_temp$class_id,
+                                     Lnames = Lnames, xmin = 60/epochSize, mode = fragmode)
+          # fragmentation values can come with a lot of decimal places
+          summary[fi:(fi + (length(frag.out) - 1))] = round(as.numeric(frag.out), digits = 6)
+          s_names[fi:(fi + (length(frag.out) - 1))] = paste0("FRAG_", names(frag.out), "_", fragmode)
+          fi = fi + length(frag.out)
+        }
+      # } else {
+      #   cosinor_coef = NULL
+      #   s_names[fi:(fi + 20)] = c("cosinor_timeOffsetHours", "cosinor_mes", 
+      #                             "cosinor_amp", "cosinor_acrophase",
+      #                             "cosinor_acrotime", "cosinor_ndays", "cosinor_R2", 
+      #                             "cosinorExt_minimum", "cosinorExt_amp", 
+      #                             "cosinorExt_alpha", "cosinorExt_beta", 
+      #                             "cosinorExt_acrotime", "cosinorExt_UpMesor",
+      #                             "cosinorExt_DownMesor", "cosinorExt_MESOR",
+      #                             "cosinorExt_ndays", "cosinorExt_F_pseudo", 
+      #                             "cosinorExt_R2", "IS", "IV", "phi")
+      #   fi = fi + 21
       }
       
       #=======================================================================
@@ -290,9 +315,9 @@ g.part6 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
           cosinor_ts = c()
         }
         save(output_part6, cosinor_ts, file = paste0(metadatadir,
-                                         ms6.out, "/", gsub(pattern = "[.]csv|[.]RData",
-                                                            replacement = "",
-                                                            x = fnames.ms5raw[i]), ".RData"))
+                                                     ms6.out, "/", gsub(pattern = "[.]csv|[.]RData",
+                                                                        replacement = "",
+                                                                        x = fnames.ms5raw[i]), ".RData"))
       }
       rm(output_part6, summary)
     }